+++ Tests/Unit/AdminPanel/TypesInformationTest.php
@@ -101,7 +101,7 @@ class TypesInformationTest extends TestCase
         ];
 
         yield 'One type is available' => [
-            '{"@context":"http://schema.org","@type":"Thing","@id":"thingyId","description":"thingy description","name":"thingy name"}',
+            '{"@context":"https://schema.org/","@type":"Thing","@id":"thingyId","description":"thingy description","name":"thingy name"}',
             [
                 [
                     '@type' => 'Thing',
@@ -113,7 +113,7 @@ class TypesInformationTest extends TestCase
         ];
 
         yield 'Two types are available' => [
-            '{"@context":"http://schema.org","@graph":[{"@type":"Action","name":"action name","url":"http://example.org/"},{"@type":"Person","@id":"personId","name":"person name","worksFor":"someone"}]}',
+            '{"@context":"https://schema.org/","@graph":[{"@type":"Action","name":"action name","url":"http://example.org/"},{"@type":"Person","@id":"personId","name":"person name","worksFor":"someone"}]}',
             [
                 [
                     '@type' => 'Action',
@@ -130,7 +130,7 @@ class TypesInformationTest extends TestCase
         ];
 
         yield 'Types are sorted alphabetically' => [
-            '{"@context":"http://schema.org","@graph":[{"@type":"Thing","name":"A thing"},{"@type":"Event","name":"An event"},{"@type":"Person","name":"A person"}]}',
+            '{"@context":"https://schema.org/","@graph":[{"@type":"Thing","name":"A thing"},{"@type":"Event","name":"An event"},{"@type":"Person","name":"A person"}]}',
             [
                 [
                     '@type' => 'Event',