+++ Tests/Unit/ViewHelpers/PropertyViewHelperTest.php
@@ -54,7 +54,7 @@ class PropertyViewHelperTest extends ViewHelperTestCase
             '<schema:type.thing>
                     <schema:property -as="image" value="http://example.org/image.png"/>
                 </schema:type.thing>',
-            \sprintf(Extension::JSONLD_TEMPLATE, '{"@context":"http://schema.org","@type":"Thing","image":"http://example.org/image.png"}'),
+            \sprintf(Extension::JSONLD_TEMPLATE, '{"@context":"https://schema.org/","@type":"Thing","image":"http://example.org/image.png"}'),
         ];
 
         yield 'Property with multiple values' => [
@@ -64,14 +64,14 @@ class PropertyViewHelperTest extends ViewHelperTestCase
                     <schema:property -as="image" value="http://example.org/image3.png"/>
                     <schema:property -as="image" value="http://example.org/image4.png"/>
                 </schema:type.thing>',
-            \sprintf(Extension::JSONLD_TEMPLATE, '{"@context":"http://schema.org","@type":"Thing","image":["http://example.org/image1.png","http://example.org/image2.png","http://example.org/image3.png","http://example.org/image4.png"]}'),
+            \sprintf(Extension::JSONLD_TEMPLATE, '{"@context":"https://schema.org/","@type":"Thing","image":["http://example.org/image1.png","http://example.org/image2.png","http://example.org/image3.png","http://example.org/image4.png"]}'),
         ];
 
         yield 'Property with value "0"' => [
             '<schema:type.thing>
                     <schema:property -as="isAccessibleForFree" value="0"/>
                 </schema:type.thing>',
-            \sprintf(Extension::JSONLD_TEMPLATE, '{"@context":"http://schema.org","@type":"Thing","isAccessibleForFree":"0"}'),
+            \sprintf(Extension::JSONLD_TEMPLATE, '{"@context":"https://schema.org/","@type":"Thing","isAccessibleForFree":"0"}'),
         ];
     }