+++ Tests/Unit/ViewHelpers/Type/ThingViewHelperTest.php
@@ -73,7 +73,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                     name="thingy name"
                     description="thingy description"
                 />',
-            '{"@context":"http://schema.org","@type":"Thing","@id":"thingyId","description":"thingy description","name":"thingy name"}',
+            '{"@context":"https://schema.org/","@type":"Thing","@id":"thingyId","description":"thingy description","name":"thingy name"}',
         ];
 
         yield 'Multiple types' => [
@@ -91,7 +91,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                     name="action name"
                     url="http://example.org/"
                 />',
-            '{"@context":"http://schema.org","@graph":[{"@type":"Thing","@id":"thingyId","description":"thingy description","name":"thingy name"},{"@type":"Person","@id":"personId","name":"person name","worksFor":"someone"},{"@type":"Action","name":"action name","url":"http://example.org/"}]}',
+            '{"@context":"https://schema.org/","@graph":[{"@type":"Thing","@id":"thingyId","description":"thingy description","name":"thingy name"},{"@type":"Person","@id":"personId","name":"person name","worksFor":"someone"},{"@type":"Action","name":"action name","url":"http://example.org/"}]}',
         ];
 
         yield 'On top level type -as is ignored' => [
@@ -99,7 +99,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                     -as="shouldBeIgnored"
                     name="as is ignored"
                 />',
-            '{"@context":"http://schema.org","@type":"Thing","name":"as is ignored"}',
+            '{"@context":"https://schema.org/","@type":"Thing","name":"as is ignored"}',
         ];
 
         yield 'Type with one child' => [
@@ -114,7 +114,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                         url="https://example.org/child"
                     />
                 </schema:type.thing>',
-            '{"@context":"http://schema.org","@type":"Thing","@id":"parentThing","name":"parent name","subjectOf":{"@type":"Person","@id":"childThing","name":"child name","url":"https://example.org/child"},"url":"http://example.org/"}',
+            '{"@context":"https://schema.org/","@type":"Thing","@id":"parentThing","name":"parent name","subjectOf":{"@type":"Person","@id":"childThing","name":"child name","url":"https://example.org/child"},"url":"http://example.org/"}',
         ];
 
         yield 'Type with multiple childs' => [
@@ -142,14 +142,14 @@ class ThingViewHelperTest extends ViewHelperTestCase
                         areaServed="TR"
                     />
                 </schema:type.organization>',
-            '{"@context":"http://schema.org","@type":"Organization","contactPoint":[{"@type":"ContactPoint","areaServed":"DE","contactType":"sales","telephone":"+49 30 123456789"},{"@type":"ContactPoint","areaServed":"PL","contactType":"sales","telephone":"+48 22 123456789"},{"@type":"ContactPoint","areaServed":"TR","contactType":"sales","telephone":"+90 212 123456789"}],"logo":"https://www.example.org/logo.png","name":"Acme Ltd.","url":"https://www.example.org/"}',
+            '{"@context":"https://schema.org/","@type":"Organization","contactPoint":[{"@type":"ContactPoint","areaServed":"DE","contactType":"sales","telephone":"+49 30 123456789"},{"@type":"ContactPoint","areaServed":"PL","contactType":"sales","telephone":"+48 22 123456789"},{"@type":"ContactPoint","areaServed":"TR","contactType":"sales","telephone":"+90 212 123456789"}],"logo":"https://www.example.org/logo.png","name":"Acme Ltd.","url":"https://www.example.org/"}',
         ];
 
         yield 'Type inside "for" loop' => [
             '<f:for each="{0: \'foo\', 1: \'bar\', 2: \'qux\'}" as="item" iteration="iterator">
                 <schema:type.thing name="{item}" identifier="{iterator.cycle}"/>
             </f:for>',
-            '{"@context":"http://schema.org","@graph":[{"@type":"Thing","identifier":"1","name":"foo"},{"@type":"Thing","identifier":"2","name":"bar"},{"@type":"Thing","identifier":"3","name":"qux"}]}',
+            '{"@context":"https://schema.org/","@graph":[{"@type":"Thing","identifier":"1","name":"foo"},{"@type":"Thing","identifier":"2","name":"bar"},{"@type":"Thing","identifier":"3","name":"qux"}]}',
         ];
 
         yield 'Type with -isMainEntityOfWebPage set to true without a WebPage' => [
@@ -165,7 +165,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                     url="https://example.org/child"
                 />
                 </schema:type.thing>',
-            '{"@context":"http://schema.org","@type":"Thing","@id":"parentThing","name":"parent name","subjectOf":{"@type":"Person","@id":"childThing","name":"child name","url":"https://example.org/child"},"url":"http://example.org/"}',
+            '{"@context":"https://schema.org/","@type":"Thing","@id":"parentThing","name":"parent name","subjectOf":{"@type":"Person","@id":"childThing","name":"child name","url":"https://example.org/child"},"url":"http://example.org/"}',
         ];
 
         yield 'Type with -isMainEntityOfWebPage set to "1" with a WebPage' => [
@@ -182,7 +182,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                     url="https://example.org/child"
                 />
             </schema:type.thing>',
-            '{"@context":"http://schema.org","@type":"WebPage","mainEntity":{"@type":"Thing","@id":"parentThing","name":"parent name","subjectOf":{"@type":"Person","@id":"childThing","name":"child name","url":"https://example.org/child"},"url":"http://example.org/"}}',
+            '{"@context":"https://schema.org/","@type":"WebPage","mainEntity":{"@type":"Thing","@id":"parentThing","name":"parent name","subjectOf":{"@type":"Person","@id":"childThing","name":"child name","url":"https://example.org/child"},"url":"http://example.org/"}}',
         ];
 
         yield 'More than on type with -isMainEntityOfWebPage set to "1" with a WebPage' => [
@@ -197,7 +197,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                     -isMainEntityOfWebPage="1"
                     name="parent name #2"
                  />',
-            '{"@context":"http://schema.org","@type":"WebPage","mainEntity":[{"@type":"Thing","@id":"parentThing#1","name":"parent name #1"},{"@type":"Thing","@id":"parentThing#2","name":"parent name #2"}]}',
+            '{"@context":"https://schema.org/","@type":"WebPage","mainEntity":[{"@type":"Thing","@id":"parentThing#1","name":"parent name #1"},{"@type":"Thing","@id":"parentThing#2","name":"parent name #2"}]}',
         ];
 
         yield 'Type with -isMainEntityOfWebPage set to "true" with a WebPage' => [
@@ -214,7 +214,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                     url="https://example.org/child"
                 />
             </schema:type.thing>',
-            '{"@context":"http://schema.org","@type":"WebPage","mainEntity":{"@type":"Thing","@id":"parentThing","name":"parent name","subjectOf":{"@type":"Person","@id":"childThing","name":"child name","url":"https://example.org/child"},"url":"http://example.org/"}}',
+            '{"@context":"https://schema.org/","@type":"WebPage","mainEntity":{"@type":"Thing","@id":"parentThing","name":"parent name","subjectOf":{"@type":"Person","@id":"childThing","name":"child name","url":"https://example.org/child"},"url":"http://example.org/"}}',
         ];
 
         yield 'Type with -isMainEntityOfWebPage set to "0" with a WebPage' => [
@@ -225,7 +225,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                 name="parent name"
                 url="http://example.org/"
             />',
-            '{"@context":"http://schema.org","@graph":[{"@type":"WebPage"},{"@type":"Thing","@id":"parentThing","name":"parent name","url":"http://example.org/"}]}',
+            '{"@context":"https://schema.org/","@graph":[{"@type":"WebPage"},{"@type":"Thing","@id":"parentThing","name":"parent name","url":"http://example.org/"}]}',
         ];
 
         yield 'Type with -isMainEntityOfWebPage set to "false" with a WebPage' => [
@@ -236,7 +236,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                 name="parent name"
                 url="http://example.org/"
             />',
-            '{"@context":"http://schema.org","@graph":[{"@type":"WebPage"},{"@type":"Thing","@id":"parentThing","name":"parent name","url":"http://example.org/"}]}',
+            '{"@context":"https://schema.org/","@graph":[{"@type":"WebPage"},{"@type":"Thing","@id":"parentThing","name":"parent name","url":"http://example.org/"}]}',
         ];
 
         yield 'Property value of 0.00 is rendered' => [
@@ -244,7 +244,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
                 price="0.00"
                 priceCurrency="EUR"
              />',
-            '{"@context":"http://schema.org","@type":"Offer","price":"0","priceCurrency":"EUR"}',
+            '{"@context":"https://schema.org/","@type":"Offer","price":"0","priceCurrency":"EUR"}',
         ];
 
         yield 'Property value of 0.01 is rendered' => [
@@ -252,21 +252,21 @@ class ThingViewHelperTest extends ViewHelperTestCase
                 price="0.01"
                 priceCurrency="EUR"
              />',
-            '{"@context":"http://schema.org","@type":"Offer","price":"0.01","priceCurrency":"EUR"}',
+            '{"@context":"https://schema.org/","@type":"Offer","price":"0.01","priceCurrency":"EUR"}',
         ];
 
-        yield 'Property value of "false" is rendered to http://schema.org/False' => [
+        yield 'Property value of "false" is rendered to https://schema.org/False' => [
             '<schema:type.event
                 isAccessibleForFree="false"
              />',
-            '{"@context":"http://schema.org","@type":"Event","isAccessibleForFree":"http://schema.org/False"}',
+            '{"@context":"https://schema.org/","@type":"Event","isAccessibleForFree":"https://schema.org/False"}',
         ];
 
-        yield 'Property value of "true" is rendered to http://schema.org/True' => [
+        yield 'Property value of "true" is rendered to https://schema.org/True' => [
             '<schema:type.event
                 isAccessibleForFree="true"
              />',
-            '{"@context":"http://schema.org","@type":"Event","isAccessibleForFree":"http://schema.org/True"}',
+            '{"@context":"https://schema.org/","@type":"Event","isAccessibleForFree":"https://schema.org/True"}',
         ];
     }
 
@@ -292,7 +292,7 @@ class ThingViewHelperTest extends ViewHelperTestCase
     public function itRecognisesAGivenSpecificType(): void
     {
         $template = '<schema:type.organization name="a corporation" -specificType="Corporation"/>';
-        $expected = \sprintf(Extension::JSONLD_TEMPLATE, '{"@context":"http://schema.org","@type":"Corporation","name":"a corporation"}');
+        $expected = \sprintf(Extension::JSONLD_TEMPLATE, '{"@context":"https://schema.org/","@type":"Corporation","name":"a corporation"}');
 
         $this->renderTemplate($template);